+++ examples/Net8.0/Example8-EtlPipeline/ConsoleColors.cs
@@ -1,0 +1,8 @@
+namespace Example8_EtlPipeline;
+
+internal class ConsoleColors
+{
+    public const string Green = "\e[32m";
+    public const string Yellow = "\e[33m";
+    public const string Reset = "\e[0m";
+}